DocumentPickerDelegate 코드 분리
